chore: update fingerprinthub
